- changed from .gui to .ui (class variables name) - added wasPlaying (metronome) to the save/load - when focus is on a SpinButton, its still possible to close a SecundaryWindow ('esc' key)

+++ sound.cpp
@@ -71,13 +71,13 @@ isPlaying_var = false;
 
 
 
-void Sound::setFrequency (int frequency)
+void Sound::setFrequency (double frequency)
 {
 normalFrequency_var = frequency;
 }
 
 
-void Sound::setStrongFrequency (int frequency)
+void Sound::setStrongFrequency (double frequency)
 {
 strongFrequency_var = frequency;
 }
